+++ posthog/src/test/java/com/posthog/PostHogTest.kt
@@ -574,13 +574,88 @@ internal class PostHogTest {
 
         sut.identify(
             "anotherDistinctId",
+        )
+
+        queueExecutor.shutdownAndAwaitTermination()
+
+        assertEquals(1, http.requestCount)
+
+        sut.close()
+    }
+
+    @Test
+    fun `captures a set event if identified`() {
+        val http = mockHttp()
+        val url = http.url("/")
+
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, reloadFeatureFlags = false, flushAt = 2)
+
+        sut.identify(
+            DISTINCT_ID,
+            userProperties = userProps,
+            userPropertiesSetOnce = userPropsOnce,
+        )
+
+        val userProps = mapOf("user1" to "theResult")
+        val userPropsOnce = mapOf("logged" to false)
+
+        sut.identify(
+            DISTINCT_ID,
             userProperties = userProps,
             userPropertiesSetOnce = userPropsOnce,
         )
 
         queueExecutor.shutdownAndAwaitTermination()
 
-        assertEquals(1, http.requestCount)
+        val request = http.takeRequest()
+
+        val content = request.body.unGzip()
+        val batch = serializer.deserialize<PostHogBatchEvent>(content.reader())
+
+        val theEvent = batch.batch.last()
+
+        assertEquals("\$set", theEvent.event)
+        assertEquals(userProps, theEvent.properties!!["\$set"])
+        assertEquals(userPropsOnce, theEvent.properties!!["\$set_once"])
+
+        sut.close()
+    }
+
+    @Test
+    fun `does not capture a set event if different user`() {
+        val http = mockHttp()
+        val url = http.url("/")
+
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, reloadFeatureFlags = false)
+
+        sut.identify(
+            DISTINCT_ID,
+            userProperties = userProps,
+            userPropertiesSetOnce = userPropsOnce,
+        )
+
+        val userProps = mapOf("user1" to "theResult")
+        val userPropsOnce = mapOf("logged" to false)
+
+        sut.identify(
+            "different user",
+            userProperties = userProps,
+            userPropertiesSetOnce = userPropsOnce,
+        )
+
+        queueExecutor.shutdownAndAwaitTermination()
+
+        queueExecutor.shutdownAndAwaitTermination()
+
+        val request = http.takeRequest()
+
+        val content = request.body.unGzip()
+        val batch = serializer.deserialize<PostHogBatchEvent>(content.reader())
+
+        val theEvent = batch.batch.last()
+
+        assertEquals(1, batch.batch.size)
+        assertEquals("\$identify", theEvent.event)
 
         sut.close()
     }