+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- fix: allow changing person properties after identify ([#205](https://github.com/PostHog/posthog-android/pull/205)) 
+
 ## 3.9.1 - 2024-11-11
 
 - recording: fix observation on multiple threads in layout/draw is not supported for compose ([#204](https://github.com/PostHog/posthog-android/pull/204))